ロールによる機能制限

+++ routes/api.php
@@ -42,28 +42,28 @@ Route::post('/management/login', [AuthorizationController::class, 'adminLogin'])
 Route::group(['middleware' => ['auth:sanctum']], function () {
 
     // 管理者アカウントの一覧取得 (ownerは自分の店舗に所属しているアカウントのみ取得可能)
-    Route::get('/management/account', [AccountManagementController::class, 'adminList']);
+    Route::middleware('adminRoleRestriction')->get('/management/account', [AccountManagementController::class, 'adminList']);
 
     // 新規管理者アカウント(owner, counter, kitchen)発行
-    Route::post('/management/account', [AccountManagementController::class, 'adminSignUp']);
+    Route::middleware('adminRoleRestriction')->post('/management/account', [AccountManagementController::class, 'adminSignUp']);
 
     // 管理者アカウント(owner, counter, kitchen)削除
-    Route::delete('/management/account/{admin_id}', [AccountManagementController::class, 'adminDelete']);
+    Route::middleware('adminRoleRestriction')->delete('/management/account/{admin_id}', [AccountManagementController::class, 'adminDelete']);
 
     // 管理者アカウント情報取得
     Route::get('/management/account/{admin_id}', [AccountManagementController::class, 'adminGet']);
 
-    // Summary: 管理者アカウント(owner, counter, kitchen)情報編集
+    // 管理者アカウント(owner, counter, kitchen)情報編集
     Route::put('/management/account/{admin_id}', [AccountManagementController::class, 'adminModify']);
 
     // メニューに料理を追加(新規作成)
-    Route::post('/management/dish/byRestaurant/{restaurant_id}', [DishManagementController::class, 'dishCreate']);
+    Route::middleware('adminRoleRestriction')->post('/management/dish/byRestaurant/{restaurant_id}', [DishManagementController::class, 'dishCreate']);
 
     // 料理メニューの一覧取得
     Route::get('/management/dish/byRestaurant/{restaurant_id}', [DishManagementController::class, 'dishList']);
 
     // 料理の削除
-    Route::delete('/management/dish/{dish_id}', [DishManagementController::class, 'dishDelete']);
+    Route::middleware('adminRoleRestriction')->delete('/management/dish/{dish_id}', [DishManagementController::class, 'dishDelete']);
 
     // 指定した料理情報取得
     Route::get('/management/dish/{dish_id}', [DishManagementController::class, 'dishGet']);
@@ -90,28 +90,28 @@ Route::group(['middleware' => ['auth:sanctum']], function () {
     Route::put('/management/order/{order_id}/checkout', [OrderManagementController::class, 'orderComplete']);
 
     // 店舗の一覧取得 (ownerは自分の店舗のみ取得可能)
-    Route::get('/management/restaurant', [RestaurantManagementController::class, 'restaurantList']);
+    Route::middleware('adminRoleRestriction')->get('/management/restaurant', [RestaurantManagementController::class, 'restaurantList']);
 
     // 新規店舗登録
-    Route::post('/management/restaurant', [RestaurantManagementController::class, 'restaurantSignUp']);
+    Route::middleware('adminRoleRestriction')->post('/management/restaurant', [RestaurantManagementController::class, 'restaurantSignUp']);
 
     // 店舗の削除
-    Route::delete('/management/restaurant/{restaurant_id}', [RestaurantManagementController::class, 'restaurantDelete']);
+    Route::middleware('adminRoleRestriction')->delete('/management/restaurant/{restaurant_id}', [RestaurantManagementController::class, 'restaurantDelete']);
 
     // 指定した店舗情報取得
     Route::get('/management/restaurant/{restaurant_id}', [RestaurantManagementController::class, 'restaurantGet']);
 
     // 店舗情報編集
-    Route::put('/management/restaurant/{restaurant_id}', [RestaurantManagementController::class, 'restaurantModify']);
+    Route::middleware('adminRoleRestriction')->put('/management/restaurant/{restaurant_id}', [RestaurantManagementController::class, 'restaurantModify']);
 
     // 座席の追加
-    Route::post('/management/seat/byRestaurant/{restaurant_id}', [SeatManagementController::class, 'seatAdd']);
+    Route::middleware('adminRoleRestriction')->post('/management/seat/byRestaurant/{restaurant_id}', [SeatManagementController::class, 'seatAdd']);
 
     // 指定した店舗の座席情報一覧を取得
     Route::get('/management/seat/byRestaurant/{restaurant_id}', [SeatManagementController::class, 'seatList']);
 
     // 座席の削除
-    Route::delete('/management/seat/{seat_id}', [SeatManagementController::class, 'seatDelete']);
+    Route::middleware('adminRoleRestriction')->delete('/management/seat/{seat_id}', [SeatManagementController::class, 'seatDelete']);
 
     // 座席の編集 (編集後、自動QRコードトークン再発行)
     Route::put('/management/seat/{seat_id}', [SeatManagementController::class, 'seatEdit']);
@@ -123,5 +123,5 @@ Route::group(['middleware' => ['auth:sanctum']], function () {
     Route::put('/management/seat/{seat_id}/refresh', [SeatManagementController::class, 'seatRefresh']);
 
     // 新規店舗の座席一括登録
-    Route::post('/management/seat/byRestaurant/{restaurant_id}/{seat_num}', [SeatManagementController::class, 'multipleSeatAdd']);
+    Route::middleware('adminRoleRestriction')->post('/management/seat/byRestaurant/{restaurant_id}/{seat_num}', [SeatManagementController::class, 'multipleSeatAdd']);
 });
